Added exciting new modules for nn and Optimisers
