+++ src/nn/mod.rs
@@ -1,0 +1,2 @@
+pub mod linear;
+pub mod model;
