+++ src/tensor/mod.rs
@@ -5,6 +5,7 @@ mod raw_tensor;
 mod rc_tensor;
 mod tensor_like;
 mod tensor_view;
+mod types;
 mod utils;
 
 pub use numeric::*;
@@ -12,3 +13,4 @@ pub use raw_tensor::*;
 pub use rc_tensor::*;
 pub use tensor_like::*;
 pub use tensor_view::*;
+pub use types::*;
