+++ src/tensor/numeric.rs
@@ -1,9 +1,12 @@
 pub use num::{One, Zero};
 use std::cmp::PartialEq;
 
-pub use std::ops::{Add, Mul, Neg};
+pub use std::ops::{Add, Mul, Neg, Sub};
 
-pub trait Numeric: Zero + One + Copy + Clone + Mul + Add + PartialEq + std::fmt::Debug {}
+pub trait Numeric:
+    Zero + One + Copy + Clone + Mul + Add + Sub + PartialEq + std::fmt::Debug
+{
+}
 // https://stackoverflow.com/questions/42381185/specifying-generic-parameter-to-belong-to-a-small-set-of-types
 macro_rules! numeric_impl {
     ($($t: ty),+) => {
